refactor: cleaned up MegaMenu components
