+++ packages/web/utils/menuIcons.ts
@@ -1,0 +1,53 @@
+import Alliances from 'assets/menuIcon/alliances.svg';
+import Asketh from 'assets/menuIcon/asketh.svg';
+import BecomeAPatron from 'assets/menuIcon/becomeapatron.svg';
+import Contribute from 'assets/menuIcon/contribute.svg';
+import Discord from 'assets/menuIcon/discord.svg';
+import Events from 'assets/menuIcon/events.svg';
+import Forum from 'assets/menuIcon/forum.svg';
+import Grants from 'assets/menuIcon/grants.svg';
+import Guilds from 'assets/menuIcon/guilds.svg';
+import Invest from 'assets/menuIcon/invest.svg';
+import Learn from 'assets/menuIcon/learn.svg';
+import MetaGameWiki from 'assets/menuIcon/metagamewiki.svg';
+import MetaRadio from 'assets/menuIcon/metaradio.svg';
+import Patrons from 'assets/menuIcon/patrons.svg';
+import Playbooks from 'assets/menuIcon/playbooks.svg';
+import Players from 'assets/menuIcon/players.svg';
+import Quests from 'assets/menuIcon/quests.svg';
+import Raids from 'assets/menuIcon/raids.svg';
+import Roles from 'assets/menuIcon/roles.svg';
+import SeedEarned from 'assets/menuIcon/seedearned.svg';
+import Seeds from 'assets/menuIcon/seeds.svg';
+import TheGreatHouses from 'assets/menuIcon/thegreathouses.svg';
+import WelcomeToMetaGame from 'assets/menuIcon/welcometometagame.svg';
+import XPEarned from 'assets/menuIcon/xpearned.svg';
+import Youtube from 'assets/menuIcon/youtube.svg';
+
+export const menuIcons: Record<string, string> = {
+  alliances: Alliances,
+  asketh: Asketh,
+  contribute: Contribute,
+  discord: Discord,
+  forum: Forum,
+  grants: Grants,
+  guilds: Guilds,
+  invest: Invest,
+  learn: Learn,
+  metagamewiki: MetaGameWiki,
+  patrons: Patrons,
+  playbooks: Playbooks,
+  players: Players,
+  quests: Quests,
+  raids: Raids,
+  roles: Roles,
+  seedearned: SeedEarned,
+  seeds: Seeds,
+  thegreathouses: TheGreatHouses,
+  welcometometagame: WelcomeToMetaGame,
+  xpearned: XPEarned,
+  youtube: Youtube,
+  metaradio: MetaRadio,
+  events: Events,
+  becomeapatron: BecomeAPatron,
+} as const;
